+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+* return `WP_Error` in `Language_Pack_API::get_language_pack()` with validation error
+* exit gracefully if `Language_Pack_API::get_language_pack()` returns `WP_Error`
 
 ####  1.1.0 / 2024-11-16
 * add API error caching/logging